refactor(protocol): consolidate simulation contracts
