+++ packages/simulation/package.json
@@ -25,6 +25,7 @@
     "@opencode-ai/core": "workspace:*",
     "@opencode-ai/ai": "workspace:*",
     "@opencode-ai/plugin": "workspace:*",
+    "@opencode-ai/protocol": "workspace:*",
     "@opencode-ai/util": "workspace:*",
     "@opentui/core": "catalog:",
     "effect": "catalog:"
